+++ java/system-test/system-test-dsl/src/main/java/sleeper/systemtest/dsl/python/PythonBulkImportDsl.java
@@ -16,10 +16,10 @@
 
 package sleeper.systemtest.dsl.python;
 
-import sleeper.systemtest.dsl.SentIngestJobsContext;
 import sleeper.systemtest.dsl.SystemTestContext;
 import sleeper.systemtest.dsl.SystemTestDrivers;
 import sleeper.systemtest.dsl.ingest.IngestByAnyQueueDriver;
+import sleeper.systemtest.dsl.ingest.SentIngestJobsContext;
 import sleeper.systemtest.dsl.util.PollWithRetriesDriver;
 import sleeper.systemtest.dsl.util.WaitForJobs;
